refactor: collapse model variants to fp32 + int4 + auto
Drop fp16 / int8 / int4f16. The fine-tune ships fp32 + int4 only and
the npm runtime now mirrors that shape — one full-precision baseline,
one quantized default. `auto` resolves to int4 on every backend
(small footprint, ~6% F1 vs fp32). Pin `variant: 'fp32'` for max
accuracy or regression.

Files updated:
- src/types/config.ts — ModelVariant = fp32 | int4 | auto
- src/defaults.ts — MANAGER_DEFAULT_VARIANT = 'int4'
- src/model-manager.ts — VARIANT_FILES drops fp16/int8/int4f16
- src/backend/variant.ts — VARIANT_TO_FILE drops same
- src/backend/{cpu,mps,cuda,rocm}-backend.ts — autoVariant = int4
- src/cli/commands/* — help strings + benchmark candidates
- test/_env.ts — gate on model_q4.onnx (was model_fp16.onnx); add
  HAS_TEST_FP32 for fp32-baseline tests
- test/{cli,nullpii,model-manager}.test.ts + test/backend/* —
  variant strings updated to int4, gating updated to HAS_TEST_ARTIFACTS
- README.md, docs/index.md, docs/guide/{getting-started,how-it-works,
  backends,eval-results}.md — variants tables + examples updated

Verification: typecheck, build, 156 tests (148 pass, 8 skipped),
lint clean.

+++ src/backend/cuda-backend.ts
@@ -8,9 +8,8 @@ import { type BackendConfig, OrtBackend, type SessionThreads } from './ort-backe
 const CUDA_CONFIG: BackendConfig = {
   name: 'cuda',
   executionProviders: [{ name: 'cuda' }, 'cpu'],
-  // CUDA: fp16 uses tensor cores (Volta+) for ~2× throughput vs fp32 on
-  // most modern GPUs, with negligible accuracy loss for this model.
-  autoVariant: 'fp16',
+  // Default to int4 (~875 MB, ~6% F1 vs fp32). Pin fp32 for accuracy.
+  autoVariant: 'int4',
 };
 
 const LINUX_HINTS = ['/dev/nvidia0', '/proc/driver/nvidia/version'] as const;
